refactor: Extract create new event type logic into standalone file
